Splitting tests into different namespaces and starting a base class for database setup and tear down
