--- flask-server/tests/test_app.py
@@ -1,1 +1,0 @@
-# Tests on requests
