+++ flask-server/tests/test_content_api.py
@@ -4,38 +4,30 @@ import sys
 sys.path.append("..")
 from app import create_app, db
 from config import TestConfig
-from routes import register_user
 
-# Still problems with initialisation
+class TestContentAPI(TestCase):
+    """Tests for content namespace routes"""
 
-class APITestCase(TestCase):
-    # This function will set up our test database
-    # Will need to rewrite this
     def setUp(self):
+        """Set up our test database and work within the context of our application"""
         self.app=create_app(TestConfig)
 
         self.client=self.app.test_client(self)
 
         with self.app.app_context():
-            db.init_app(self.app)
 
             db.create_all()
 
-
-    def test_home_response(self):
-        home_response = self.client.get('/')
-        result = home_response.json
-        expected = {"message":"hello"}
+    def test_hello_world(self):
+        """Test the hello world route"""
+        hello_response = self.client.get("/content/hello")
+        result = hello_response.json
+        expected = {"message": "Hello world!"}
 
         self.assertEqual(expected, result)
 
-    
-    def test_register_user_success(self):
-        pass
-
-
-    # This function will remove everything from our test database
     def tearDown(self):
+        """Destroy all the instances created for testing, remove sessions and drop tables"""
         with self.app.app_context():
             db.session.remove()
             db.drop_all()
